Global Respawn Manager
This is what I came up with in a short period of time, im not sure of the best way to handle this in terms of performance.

It runs great in my limited tests,  not sure how it would scale with hundreds of mobs killed in a short period of time.

Dictionary may also not be the best collection here in terms of performance, but from a readability perspective i went with it to at least test the concept.