Npc Health/MP Fix + Server CPU Utilization
Health data now loaded and set in template.

As a side effect of this change CPU usage of the server has decreased by 80%.  Now that there are not 46,000 regen tasks running for each L2Monster that spanwed with incorrect hp values.

Ultimately CharStatus needs to know what type of character it is, as the BaseHpMax calculation is really based on players, not NPC's